+++ src/client/hooks/useDeeplink.hook.ts
@@ -1,8 +1,7 @@
 import { useLocation } from 'react-router-dom';
-import { useLocalStorage, removeLocalStorageKey } from './storage.hook';
-import { isPrivateRoute } from '../../universal/helpers';
 import { AppRoutes } from '../../universal/config';
-import { useEffect } from 'react';
+import { isPrivateRoute } from '../config/routes';
+import { removeLocalStorageKey, useLocalStorage } from './storage.hook';
 
 export const ROUTE_ENTRY_KEY = 'RouteEntry';
