+++ src/client/utils/utils.ts
@@ -1,0 +1,26 @@
+import { KeyboardEvent, MouseEvent } from 'react';
+
+// Repeating conditions for accessible keyboard event
+export function withKeyPress<T>(fn: Function, keyName: string = 'enter') {
+  return (event: KeyboardEvent<T> | MouseEvent<T>) => {
+    if (!('key' in event) || event.key.toLowerCase() === keyName) {
+      fn(event);
+    }
+  };
+}
+
+/**
+ * Sloppy determination if given url points to a page of the application
+ * @param url string
+ */
+export function isInteralUrl(url: string) {
+  return (
+    url.match(/mijn\.(acc\.)?amsterdam.nl/) !== null ||
+    url.startsWith('/') ||
+    !url.startsWith('http')
+  );
+}
+
+export function isExternalUrl(url: string) {
+  return !isInteralUrl(url);
+}
