+++ src/universal/helpers/utils.ts
@@ -1,7 +1,4 @@
-import { KeyboardEvent, MouseEvent } from 'react';
-import { matchPath } from 'react-router-dom';
 import { differenceInMonths } from 'date-fns';
-import { PrivateRoutes } from '../../client/config/routes';
 
 // https://github.com/Microsoft/TypeScript/issues/21826#issuecomment-479851685
 export const entries = Object.entries as <T>(
@@ -9,42 +6,6 @@ export const entries = Object.entries as <T>(
 ) => [Extract<keyof T, string>, T[keyof T]][];
 
 export const keys = Object.keys as <T>(o: T) => (keyof T)[];
-
-// Repeating conditions for accessible keyboard event
-export function withKeyPress<T>(fn: Function, keyName: string = 'enter') {
-  return (event: KeyboardEvent<T> | MouseEvent<T>) => {
-    if (!('key' in event) || event.key.toLowerCase() === keyName) {
-      fn(event);
-    }
-  };
-}
-
-/**
- * Sloppy determination if given url points to a page of the application
- * @param url string
- */
-export function isInteralUrl(url: string) {
-  return (
-    url.match(/mijn\.(acc\.)?amsterdam.nl/) !== null ||
-    url.startsWith('/') ||
-    !url.startsWith('http')
-  );
-}
-
-export function isPrivateRoute(pathname: string) {
-  return PrivateRoutes.some(
-    (path) =>
-      !!matchPath(pathname, {
-        path,
-        exact: true,
-        strict: true,
-      })
-  );
-}
-
-export function isExternalUrl(url: string) {
-  return !isInteralUrl(url);
-}
 
 export function range(a: number, b: number) {
   return Array.from(
